requirements.txt
git+https://github.com/kshubham506/Telethon@b1984615afc684e1182ff82674dcfc0a43ed5cd9
asyncio==3.4.3
requests==2.25.1
python-binance==1.0.12
colorama==0.4.4
pyrebase4==4.5.0
numpy==1.21.2
python-dotenv==0.20.0
munch